CI: Check C# enums are in sync with latest kernel

+++ CI/azure/check_kernel.sh
@@ -5,6 +5,7 @@ KERNEL_TYPES="/tmp/mainline_types.h"
 KERNEL_MODIFIER="/tmp/modifier.c"
 IIOH="./include/iio/iio.h"
 CHANNELC="./channel.c"
+CHANNELC_SHARP="./bindings/csharp/Channel.cs"
 
 if [ ! -f ${IIOH} ] ; then
 	echo can not find ${IIOH}
@@ -16,13 +17,19 @@ if [ ! -f ${CHANNELC} ] ; then
 	exit 1
 fi
 
+if [ ! -f ${CHANNELC_SHARP} ] ; then
+	echo can not find ${CHANNELC_SHARP}
+	exit 1
+fi
+
 rm -f ${KERNEL_TYPES} ${KERNEL_MODIFIER}
 wget -O ${KERNEL_TYPES} https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/plain/include/uapi/linux/iio/types.h
 wget -O ${KERNEL_MODIFIER} https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/plain/drivers/iio/industrialio-core.c
 
+iio_groups=("iio_chan_type" "iio_modifier" "iio_event_type" "iio_event_direction")
 ret=0
 
-for enum in iio_chan_type iio_modifier iio_event_type iio_event_direction
+for enum in "${iio_groups[@]}";
 do
 	echo looking for ${enum}
 	rm -f /tmp/kernel_${enum} /tmp/libiio_${enum}
@@ -82,6 +89,34 @@ while IFS="" read -r p ; do
 	fi
 done < /tmp/kernel_modifier
 
+# Cleanup up leading tabs & spaces and trailing commas
+for group in "${iio_groups[@]}";
+do
+    sed -i -e 's/^[ \t]*//' -e 's/,$//' "/tmp/kernel_${group}"
+done
+
+echo Checking C# bindings
+
+csharp_enums=("ChannelType" "ChannelModifier")
+for i in {0..1}
+do
+	echo "looking for ${csharp_enums[i]}"
+	sed "0,/^[[:space:]]*public enum ${csharp_enums[i]}/d" ${CHANNELC_SHARP} | \
+		sed -n '/}/q;p' | sed '1{/{/d}' | sed -e 's/^[ \t]*//' -e 's/,$//' | \
+		grep -v IIO_CHAN_TYPE_UNKNOWN > "/tmp/libiio_csharp_${csharp_enums[i]}"
+
+	echo "Differences in ${csharp_enums[i]}"
+	set +e
+	diff -u -w "/tmp/libiio_csharp_${csharp_enums[i]}" "/tmp/kernel_${iio_groups[i]}"
+	count=$(diff -u -w  "/tmp/libiio_csharp_${csharp_enums[i]}" "/tmp/kernel_${iio_groups[i]}" | wc -l)
+	set -e
+	if [ "$count" -ne "0" ] ; then
+		ret=1
+		echo "difference between upstream kernel types.h and Channels.cs in ${csharp_enums[i]}"
+	else
+		echo none
+	fi
+done
 
 rm -f /tmp/kernel_modifier /tmp/libiio_iio_modifier /tmp/libiio_iio_chan_type ${KERNEL_TYPES} ${KERNEL_MODIFIER}
 exit $ret
